Trabalho 2/requirements.txt
asarPy==1.0.1
pycryptodome==3.23.0
